proxy and net.fetch
