+++ packages/insomnia/src/main.development.ts
@@ -17,6 +17,7 @@ import { registergRPCHandlers } from './main/ipc/grpc';
 import { registerMainHandlers } from './main/ipc/main';
 import { registerCurlHandlers } from './main/network/curl';
 import { registerWebSocketHandlers } from './main/network/websocket';
+import { watchProxySettings } from './main/proxy';
 import { initializeSentry, sentryWatchAnalyticsEnabled } from './main/sentry';
 import { checkIfRestartNeeded } from './main/squirrel-startup';
 import * as updates from './main/updates';
@@ -91,6 +92,7 @@ app.on('ready', async () => {
   await database.init(models.types());
   await _createModelInstances();
   sentryWatchAnalyticsEnabled();
+  watchProxySettings();
   windowUtils.init();
   await _launchApp();
